fix upgrade journey tests, dump log on replication failure

+++ replication_tunable_consistency.sh
@@ -58,31 +58,46 @@ wait_weaviate 8082
 docker run --network host -v "$PWD/workdir/data.json:/workdir/data.json" --name importer -t importer
 
 # Read objects with consistency level ONE
-docker run --network host -v "$PWD/workdir/:/workdir/data" --name cluster_healthy -t cluster_healthy
+if !docker run --network host -v "$PWD/workdir/:/workdir/data" --name cluster_healthy -t cluster_healthy; then
+  docker-compose -f apps/weaviate/docker-compose-replication.yml logs weaviate-node-1 weaviate-node-2 weaviate-node-3
+  exit 1
+fi
 
 # PATCH objects with one node down, consistency level QUORUM
 echo "Killing node 3"
 docker-compose -f apps/weaviate/docker-compose-replication.yml kill weaviate-node-3
-docker run --network host -v "$PWD/workdir/data.json:/workdir/data.json" --name patcher -t patcher
+if !docker run --network host -v "$PWD/workdir/data.json:/workdir/data.json" --name patcher -t patcher; then
+  docker-compose -f apps/weaviate/docker-compose-replication.yml logs weaviate-node-1 weaviate-node-2 weaviate-node-3
+  exit 1
+fi
 
 # Restart dead node, read objects with consistency level QUORUM
 echo "Restart node 3"
 docker-compose -f apps/weaviate/docker-compose-replication.yml up -d weaviate-node-3
 wait_weaviate 8082
-docker run --network host -v "$PWD/workdir/:/workdir/data" --name cluster_one_node_down -t cluster_one_node_down
+if !docker run --network host -v "$PWD/workdir/:/workdir/data" --name cluster_one_node_down -t cluster_one_node_down; then
+  docker-compose -f apps/weaviate/docker-compose-replication.yml logs weaviate-node-1 weaviate-node-2 weaviate-node-3
+  exit 1
+fi
 
 # PUT objects with only one node remaining, consistency level ONE
 echo "Killing node 2"
 docker-compose -f apps/weaviate/docker-compose-replication.yml kill weaviate-node-2
 echo "Killing node 3"
 docker-compose -f apps/weaviate/docker-compose-replication.yml kill weaviate-node-3
-docker run --network host -v "$PWD/workdir/data.json:/workdir/data.json" --name updater -t updater
+if !docker run --network host -v "$PWD/workdir/data.json:/workdir/data.json" --name updater -t updater; then
+  docker-compose -f apps/weaviate/docker-compose-replication.yml logs weaviate-node-1 weaviate-node-2 weaviate-node-3
+  exit 1
+fi
 
 # Restart dead nodes, read objects with consistency level ALL
 docker-compose -f apps/weaviate/docker-compose-replication.yml up -d weaviate-node-2
 wait_weaviate 8081
 docker-compose -f apps/weaviate/docker-compose-replication.yml up -d weaviate-node-3
 wait_weaviate 8082
-docker run --network host -v "$PWD/workdir/:/workdir/data" --name cluster_one_node_remaining -t cluster_one_node_remaining
+if !docker run --network host -v "$PWD/workdir/:/workdir/data" --name cluster_one_node_remaining -t cluster_one_node_remaining; then
+  docker-compose -f apps/weaviate/docker-compose-replication.yml logs weaviate-node-1 weaviate-node-2 weaviate-node-3
+  exit 1
+fi
 
 echo "Success!"
